separate BasePage, moved some functions to step
now BasePage removed and separate to BankManagerPage, AddCustomerPage, CustomersPage. Functions getClientForDeletion, getCustomersFirstNames, getCustomersLastNames was reworked to Steps
