+++ src/test/resources/test_suite.xml
@@ -1,4 +1,4 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.1.dtd" >
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.1.dtd">
 <suite name="BankingProject Test Suite" parallel="tests" thread-count="3">
     <test name="Add Customer Test">
         <classes>
@@ -9,6 +9,7 @@
             </class>
         </classes>
     </test>
+
     <test name="Sort Customers Table Test">
         <classes>
             <class name="tests.BaseTest">
@@ -18,6 +19,7 @@
             </class>
         </classes>
     </test>
+
     <test name="Deletion Customer From Customers Table Test">
         <classes>
             <class name="tests.BaseTest">
